+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - changed: (Zcash) Updated address explorer
 - changed: (Tron) Special case the `usdt-trc20` uri prefix
 - fixed: Prevent fatal error reporting for missing txlist json file
+- fixed: (Zcash) Use additional insufficient funds check before sending amount to synchronizer
 
 ## 4.14.0 (2024-07-19)